Fix common issues with running gvim on windows
Default settings for gvim on windows have many issues, such as encoding
or the annoying tendency to create a lot of persisting files in your
working directories. This commit fixes those issues, and introduces a
few common mappings and settings.
This commit also sets appearance configuration directly in the repo.