Fix Windows compatibility and packpath issues
The vimrc file is now responsible for adding the repo to the user's
packpath (this should work on Linux and Windows, no matter where the
repo is cloned, as long as the vimrc file is sourced). The
fnameescape() helps deal with special characters and spaces in paths.
Gui options are now in a separate file, they are sourced only when
`has("gui_running")` is true.