cleanup language servers config.
The time has come. Almost all plugins have been rotated at least once
and many options have been tried. The user manual has been read time and
time again. And during this time, not a line of actually usefull code
was written. This config is fine, and probably most of it will never be
used anyways. It is time to go and do some coding before we come back to
see what can be improved.